Added scaling of Beta by local connectivity rather than population. This turns the understanding of beta into a multiplication of the proportion of infected individuals in a local neighbourhood, rather than a linear multiplication of infected individuals in a local neighbourhood.
This is consistent with the idea that an individual must split their 
time in a day amongst their neighbourhood and the quality of time/amount 
of time spent with any given neighbour decreases as neighbourhood 
size/connectivity increases. This is the procedure used more frequently, 
than what we had before, in literature.

In the future I may scale this to a degree to allow some highly 
connected individuals to have more quality time available and be more 
likely to become super spreaders - activity that was better caught 
before I implemented this change.