Created a simple branching process (bpMain) and extended it to allow distributed infection times, and thinning to allow time dependent features using thinning such as saturation of susceptible individuals in a population.
The current thinning process will produce an equivalent result to any of 
discrete_branch! (with small time step ≈0.1days), firstReact_branch! and 
nextReact_branch! but is very in-efficient. 

A few other heuristics will be created to do this thinning better. 

A thinning tree (based on two structs, Tree and Node) and dependency 
tree (based on a graph type) can be produced using the population_df 
dataframe within bpMain!. The two functions should produce equivalent 
results with the same inputs. These will be useful for efficient 
thinning