Deleted most unnecessary comments and tidied some descriptions from branchingProcesses.jl
Added reuse of rows in population_df / thinned segements of the 
infection tree within the function ThinTree to improve performance. This 
is allowable because, aside from the infection time of the node that is 
the thinned, all children of the node and their children etc. are not 
dependent on attributes of their parent, so reuse is allowed.

Rerun verification graphs to demonstrate the code is working