Added a two way plot for verifying x1 vs x2 for shared y with an option of overplotting using dots or lines.
Added a stopping condition in branchingProcess() for letting model.t_max 
represent the maximum generation number when infections are not 
distributed in time. 

Added stopping condition in nextReact_branch! for if the heap storing 
events is empty. Technically this should be caught by the earlier check 
model.state_totals[2] != 0, but it was breaking anyway. This may be a 
bug

Added new tests (#15 for purely generation-based and #16 for time 
dependence) for comparing simple branching processes to an exponential 
solution. #15 is exactly equivalent, while #16 is currently a good 
approximation for the initial part of the simple branch simulation.