Added support for plotting many realisations of cumulative infections in a new function branchSideBySideVerifyPlot (within BranchVerifySoln.jl). Worked out how to use Seaborn.lineplot() to plot the mean and SD of these realisations.
Updated labels to be more appropriate to the plotting function they are 
within.

Added function for quickly returning a single quadratic spline 
interpolation.

Added newline within verifySolutions after each test to make it easier 
to read.

Added new tests (16 and 18) which are the same as 15 and 17 (prev 16) 
but compare turning the heterogeneity of R_i on and off within the  
model. Plots multiple realisations of this rather than just the mean 
(which is found to be the same).

Proper geometric series comparison found for test 17 (prev 16). 
Demonstrated equivalence to averaged time distributed branching process