Added Tests 19-21 in main_branchingProcess.jl Demonstrates the difference between using heterogeneous and homogeneous reproduction numbers for Mike et. al. style branching processes.
Test 19 and 20 show equivalent output, further demonstrating the 
equivalence of the thinned simple branching process and the next 
reaction style branching process. 

Test 21 shows the same thing as 20 + also the affect of turning 
isolation on and off.

I have also run the next react branching process model to estimate the 
epidemic curve for the August 2021 Covid Delta outbreak in New Zealand. 
5000 realisations of this with heterogeneous and homogeneous. I have 
created a new plotting function that plots the epidemic curves on the 
left and violin plots of infection counts at times of interest. Summary 
statistics at these times of interest are outputted to the REPL.