New CSVs, new conditioned ensembles, new outputs from ensembles.
Fixed the using statement in Benchmarking.jl

New tests in BPVerifySolutions.jl - trying to make the discrete and next 
BP detected amounts equivalent when using a timestep of 1 day

Updated ylabels of number of individuals in state.

Added tracking of detected cases to discrete method.

Created contact tracing for next reaction method. 

Created function for returning the Reff of a model when within an 
intervention state.